ref: move functions to analyse calls in a specific module
